# docs/swagger.yaml
definitions:
  entity.ChatMessageResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      doc_id:
        type: string
      id:
        type: integer
      role:
        type: string
      user_id:
        type: integer
    type: object
  entity.ChatRequest:
    properties:
      doc_id:
        type: string
      message:
        type: string
      user_id:
        type: integer
    required:
    - message
    - user_id
    type: object
  entity.ChatResponse:
    properties:
      reply:
        type: string
    type: object
  entity.CreatedUserResponse:
    properties:
      user_id:
        type: integer
    type: object
  entity.DocumentDeleteResponse:
    properties:
      message:
        type: string
    type: object
  entity.DocumentResponse:
    properties:
      content_type:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      id:
        type: string
      summary:
        type: string
      uploaded_at:
        type: string
      user_id:
        type: integer
    type: object
  entity.ErrorMessage:
    properties:
      message:
        type: string
    type: object
  entity.HealthMetric:
    properties:
      metric_type:
        type: string
      metric_value:
        type: string
    type: object
  entity.HealthMetricsRequest:
    properties:
      location:
        $ref: '#/definitions/entity.Location'
      metrics:
        items:
          $ref: '#/definitions/entity.HealthMetric'
        type: array
      user_id:
        type: integer
    type: object
  entity.Location:
    properties:
      latitude:
        example: 55.755825
        type: number
      longitude:
        example: 37.617396
        type: number
    type: object
  entity.NearestHospitalsRequest:
    properties:
      latitude:
        example: 55.755825
        type: number
      longitude:
        example: 37.617396
        type: number
      radius:
        description: in meters
        example: 5000
        type: integer
    type: object
  entity.NearestUsersRequest:
    properties:
      latitude:
        example: 55.755825
        type: number
      longitude:
        example: 37.617396
        type: number
      radius:
        description: in meters
        example: 5000
        type: integer
      user_id:
        type: integer
    type: object
  entity.NearestUsersResponse:
    properties:
      location:
        $ref: '#/definitions/entity.Location'
      user_id:
        type: integer
    type: object
  entity.RecommendationResponse:
    properties:
      recommendation:
        type: string
    type: object
  entity.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  entity.UserRegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: API for Dockify backend.
  title: Dockify Backend API
  version: "1.0"
paths:
  /api/v1/chat:
    delete:
      description: Delete chat history for a user, optionally filtered by document
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      - description: Document ID
        in: query
        name: doc_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Clear chat history
      tags:
      - Chat
    get:
      description: Returns chat message history for a user, optionally filtered by
        document
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      - description: Document ID
        in: query
        name: doc_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ChatMessageResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Get chat history
      tags:
      - Chat
    post:
      consumes:
      - application/json
      description: Send a message and get a response from the AI assistant
      parameters:
      - description: Chat message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Send a chat message
      tags:
      - Chat
  /api/v1/chat/stream:
    post:
      consumes:
      - application/json
      description: Send a message and receive a streaming SSE response from the AI
        assistant
      parameters:
      - description: Chat message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.ChatRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Send a chat message (streaming)
      tags:
      - Chat
  /api/v1/documents:
    get:
      consumes:
      - application/json
      description: Returns all documents uploaded by a user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.DocumentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: List user documents
      tags:
      - Documents
  /api/v1/documents/{id}:
    delete:
      description: Delete a document by ID. Validates user ownership.
      parameters:
      - description: Document UUID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DocumentDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Delete a document
      tags:
      - Documents
  /api/v1/documents/{id}/download:
    get:
      description: Download a file by document UUID
      parameters:
      - description: Document UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: file content
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Download a document
      tags:
      - Documents
  /api/v1/documents/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file (PDF, Word, image) for a user. Max 10 MB.
      parameters:
      - description: User ID
        in: formData
        name: user_id
        required: true
        type: integer
      - description: Document file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Upload a medical document
      tags:
      - Documents
  /api/v1/hospitals/nearest:
    post:
      consumes:
      - application/json
      description: Returns a list of nearest hospitals to the provided location
      parameters:
      - description: Nearest hospitals request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.NearestHospitalsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Location'
            type: array
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Get Nearest Hospitals
      tags:
      - Hospitals
  /api/v1/location/nearest:
    post:
      consumes:
      - application/json
      description: Retrieves users nearest to given coordinates within a radius.
      parameters:
      - description: Nearest users request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.NearestUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.NearestUsersResponse'
            type: array
        "204":
          description: no content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Get nearest users
      tags:
      - location
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return user information
      parameters:
      - description: User login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: authenticated user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "401":
          description: invalid email or password
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: User login
      tags:
      - User
  /api/v1/metrics:
    get:
      consumes:
      - application/json
      description: Retrieve health metrics for a given user by query parameter user_id
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: list of health metrics
          schema:
            items:
              type: object
            type: array
        "400":
          description: invalid user_id or missing parameter
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: failed to get health metrics
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Get health metrics
      tags:
      - Metrics
    post:
      consumes:
      - application/json
      description: Create health metrics for a user
      parameters:
      - description: Health metrics payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.HealthMetricsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: status message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: failed to create health metrics
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Create health metrics
      tags:
      - Metrics
  /api/v1/recommendation:
    get:
      consumes:
      - application/json
      description: Returns a recommendation string based on user metrics
      parameters:
      - description: User ID for fetching metrics
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.RecommendationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Get Recommendation
      tags:
      - Recommendation
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created user id
          schema:
            $ref: '#/definitions/entity.CreatedUserResponse'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
        "500":
          description: failed to register user
          schema:
            $ref: '#/definitions/entity.ErrorMessage'
      summary: Register a new user
      tags:
      - User
  /health:
    get:
      description: Returns the live status of the service
      produces:
      - application/json
      responses:
        "200":
          description: health
          schema:
            type: string
      summary: Health Check (Live)
      tags:
      - Health
schemes:
- http
- https
swagger: "2.0"